Copy the wrapper and *-unwrapped idiom from upstream nixpkgs
This change also moves the `isPatchedForGlobalCache` property to the
yarn package instead of the wrapper, and pass the value through the
wrapper. This is closer to what's actually happening. It does require
anyone providing a pre-built yarn binary that has been patched to wrap
it in a derivation with the appropriate passthrough, which is a minor
downside for an unlikely use case.